Refactor: Move application interfaces to Common/Interfaces
Moved `INotificationService` and `IBatchJobProcessor` from `backend/Valora.Application/Services/` to `backend/Valora.Application/Common/Interfaces/` to adhere to the Clean Architecture standard defined in the AGENTS.md document. Updated namespaces in all dependent files and verified the fix via full build and unit/integration tests.
