+++ backend/Valora.Api/Endpoints/NotificationEndpoints.cs
@@ -1,5 +1,6 @@
 using System.Security.Claims;
 using Microsoft.AspNetCore.Mvc;
+using Valora.Application.Common.Interfaces;
 using Valora.Application.Services;
 using Valora.Domain.Enums;
